Fix landing page test
Longer term this will probably want to be changed.

The difference between the actual text output and expected was
whitespace. This shouldn't affect the rendered output.

However, it would be good to understand the reason for the difference
(the returned string isn't the same as the file contents, which is
confusing). Also, the test writes to a tracked file in the repo, it
would be good to us a tmp file.